refactor: change target modules usage to map function
